Clean and move some classes
